# AurigaMail/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Espone servizi REST per le mail",
    "version" : "v1",
    "title" : "Servizi di AurigaMail",
    "termsOfService" : "",
    "contact" : {
      "name" : "API Support",
      "url" : "http://www.eng.it/",
      "email" : "Ugo.PiconeChiodo@eng.it"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "",
  "basePath" : "/AurigaMail/rest",
...
  "paths" : {
...
    "/email/noAttachment/sending" : {
      "post" : {
        "tags" : [ "Invio" ],
        "summary" : "SERVIZIO PER L'INVIO SENZA ALLEGATI.\nSpedisce una email senza allegati e salva le informazioni nel sistema. (SpedisciESalvaEmail)",
        "description" : "E' il servizio REST per testare l'invio senza allegati da UI Swagger. E' anche richiamabile con un client qualsiasi. <br/>E' stato messo a disposizione principalmente per documentare l'XML con cui valorizzare il parametro 'data'.<br/>Questo servizio ha il vantaggio di essere fruibile con richieste senza 'multipart/*'.",
        "operationId" : "sendAndSave",
        "consumes" : [ "application/xml" ],
        "produces" : [ "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "data",
          "description" : "XML richiestaSpedisciESalvaEmail",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SendAndSaveNewRequest"
          }
        }, {
          "name" : "token",
          "in" : "header",
          "description" : "Token di Autenticazione ottenibile chiamando il servizio di Login",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SendAndSaveNewResponse"
            }
          },
          "400" : {
            "description" : "Operazione interrotta a causa di input non valido."
          },
          "403" : {
            "description" : "Operazione non permessa."
          },
          "500" : {
            "description" : "Operazione interrotta a causa di un errore imprevisto."
          }
        }
      }
    },
    "/email/sending" : {
      "post" : {
        "tags" : [ "Invio" ],
        "summary" : "SERVIZIO ESPOSTO DI RIFERIMENTO.\nSpedisce una email con eventuali allegati e salva le informazioni nel sistema.(SpedisciESalvaEmail)",
        "description" : "E' il servizio REST di riferimento richiamabile con un client qualsiasi.<br/><b>Per chiamare questo servizio da UI Swagger (conviene) usare il servizio equivalente su '/email/sendingSWA'.</b><br/>Il parametro 'data' deve essere valorizzato con una stringa XML la cui struttura è documentata nel servizio su '/email/noAttachment/sending'.<br/>In realtà il tipo di dato per il parametro 'data' non è 'string' bensì un 'object' istanza di SendAndSaveNewRequest.<br/>Per un servizio così definito, purtroppo, la libreria client Swagger non è ancora in grado di produrre un file descrittore del servizio esaustivo.<br/>Di seguito un esempio di un messaggio HTTP semplificato che questo servizio prende in carico:<br/><br/>Content-Type: multipart/form-data; boundary=Boundary_1_511262261_1369143433608<br/><br/>--Boundary_1_511262261_1369143433608<br/>Content-Disposition: form-data; name=\"data\"<br/>Content-Type: application/xml<br/><br/><br/>--Boundary_1_511262261_1369143433608<br/>Content-Disposition: form-data; name=\"file\"; filename=\"aaa.pdf\"<br/>Content-Type: application/pdf<br/><br/><br/>--Boundary_1_511262261_1369143433608<br/>Content-Disposition: form-data; name=\"file\"; filename=\"bbb.txt\"<br/>Content-Type: text/plain<br/><br/><br/>--Boundary_1_511262261_1369143433608--<br/>",
        "operationId" : "sendAndSave",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/xml" ],
        "parameters" : [ {
          "name" : "file",
          "in" : "formData",
          "description" : "Files eventuali da allegare all'email",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "file"
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "token",
          "in" : "header",
          "description" : "Token di Autenticazione ottenibile chiamando il servizio di Login",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "data",
          "in" : "formData",
          "description" : "XML richiestaSpedisciESalvaEmail",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SendAndSaveNewResponse"
            }
          },
          "400" : {
            "description" : "Operazione interrotta a causa di input non valido."
          },
          "403" : {
            "description" : "Operazione non permessa."
          },
          "500" : {
            "description" : "Operazione interrotta a causa di un errore imprevisto."
          }
        }
      }
    },
    "/email/sendingSWA" : {
      "post" : {
        "tags" : [ "Invio" ],
        "summary" : "SERVIZIO AD USO ESCLUSIVO DI UI SWAGGER.\nSpedisce una email con eventuali allegati e salva le informazioni nel sistema. (SpedisciESalvaEmail)",
        "description" : "E' il servizio REST per testare l'invio con allegati da UI Swagger.<br/>Il parametro 'data' deve essere valorizzato con una stringa XML la cui struttura è documentata nel servizio su '/email/noAttachment/sending'.<br/>Conviene fare copia e incolla sul campo testo relativo.",
        "operationId" : "sendAndSave",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/xml" ],
        "parameters" : [ {
          "name" : "token",
          "in" : "header",
          "description" : "Token di Autenticazione ottenibile chiamando il servizio di Login",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "data",
          "in" : "formData",
          "description" : "XML richiestaSpedisciESalvaEmail",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "file1",
          "in" : "formData",
          "description" : "Allegato uno",
          "required" : false,
          "type" : "file"
        }, {
          "name" : "file2",
          "in" : "formData",
          "description" : "Allegato due",
          "required" : false,
          "type" : "file"
        }, {
          "name" : "file3",
          "in" : "formData",
          "description" : "Allegato tre",
          "required" : false,
          "type" : "file"
        }, {
          "name" : "file4",
          "in" : "formData",
          "description" : "Allegato quattro",
          "required" : false,
          "type" : "file"
        }, {
          "name" : "file5",
          "in" : "formData",
          "description" : "Allegato cinque",
          "required" : false,
          "type" : "file"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SendAndSaveNewResponse"
            }
          },
          "400" : {
            "description" : "Operazione interrotta a causa di input non valido."
          },
          "403" : {
            "description" : "Operazione non permessa."
          },
          "500" : {
            "description" : "Operazione interrotta a causa di un errore imprevisto."
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "EmailMessage" : {
      "type" : "object",
      "required" : [ "addressFrom" ],
      "properties" : {
        "addressFrom" : {
          "type" : "string",
          "example" : "nomeutente@dominio.it",
          "xml" : {
            "name" : "mittente"
          },
          "description" : "Indirizzo di posta elettronica del mittente"
        },
        "aliasAddressFrom" : {
          "type" : "string",
          "example" : "alias mittente",
          "xml" : {
            "name" : "aliasMittente"
          },
          "description" : "Nome logico che identifica il mittente"
        },
        "subject" : {
          "type" : "string",
          "example" : "breve descrizione del contenuto del messaggio",
          "xml" : {
            "name" : "oggetto"
          },
          "description" : "Dovrebbe aiutare il destinatario a capire il contenuto del messaggio"
        },
        "body" : {
          "type" : "string",
          "example" : "contenuto del messaggio",
          "xml" : {
            "name" : "testo"
          },
          "description" : "Contenuto informativo che il mittente vuol comunicare ai destinatari"
        }
      },
      "description" : "Un messaggio email",
      "xml" : {
        "name" : "MessaggioEmail"
      }
    },
...
    "EmailReferences" : {
      "type" : "object",
      "properties" : {
        "destinatari" : {
          "description" : "Indirizzi di posta elettronica dei destinatari dell'email.",
          "$ref" : "#/definitions/RecipientsAddresses"
        },
        "id" : {
          "type" : "string",
          "example" : "5fdae2e44e4f4e99a0c90df21e35a8dc",
          "description" : "Identificativo univoco dell'email nel sistema."
        },
        "reference" : {
          "type" : "string",
          "example" : "[FS@MAILARCHIVIATE2]/2020/4/16/1/20200416151851191203847814591582879",
          "xml" : {
            "name" : "riferimento"
          },
          "description" : "Percorso al file EML nel repository."
        },
        "messageId" : {
          "type" : "string",
          "example" : "002201d616e7$5b290640$117b12c0$@dominio.it",
          "xml" : {
            "name" : "idMessaggio"
          },
          "description" : "Identificativo del messaggio."
        }
      },
      "description" : "Riferimenti all'email elaborata nel sistema.",
      "xml" : {
        "name" : "RiferimentiEmail"
      }
    },
...
    "FileInfo" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "nome.pdf",
          "xml" : {
            "name" : "Nome"
          },
          "description" : "Nome originale del file."
        },
        "size" : {
          "type" : "string",
          "example" : "367,12 KB",
          "xml" : {
            "name" : "Dimensione"
          },
          "description" : "Dimensione del file."
        }
      },
      "description" : "Parametri essenziali di un file",
      "xml" : {
        "name" : "FileInfo"
      }
    },
...
    "RecipientsAddresses" : {
      "type" : "object",
      "properties" : {
        "addressesTo" : {
          "type" : "array",
          "xml" : {
            "name" : "a"
          },
          "description" : "Indirizzi di posta elettronica dei destinatari principali",
          "items" : {
            "type" : "string",
            "xml" : {
              "name" : "a"
            }
          }
        },
        "addressesCc" : {
          "type" : "array",
          "xml" : {
            "name" : "cc"
          },
          "description" : "Indirizzi di posta elettronica dei destinatari in copia conoscenza",
          "items" : {
            "type" : "string",
            "xml" : {
              "name" : "cc"
            }
          }
        },
        "addressesBcc" : {
          "type" : "array",
          "xml" : {
            "name" : "ccn"
          },
          "description" : "Indirizzi di posta elettronica dei destinatari in copia conoscenza nascosta",
          "items" : {
            "type" : "string",
            "xml" : {
              "name" : "ccn"
            }
          }
        }
      },
      "description" : "Destinatari dell'email",
      "xml" : {
        "name" : "Destinatari"
      }
    },
...
    "SendAndSaveNewRequest" : {
      "type" : "object",
      "required" : [ "destinatari", "messaggioEmail" ],
      "properties" : {
        "messaggioEmail" : {
          "$ref" : "#/definitions/EmailMessage"
        },
        "destinatari" : {
          "description" : "Indirizzi di posta elettronica dei destinatari dell'email.",
          "$ref" : "#/definitions/RecipientsAddresses"
        },
        "html" : {
          "type" : "boolean",
          "example" : true,
          "xml" : {
            "name" : "htmlFlag"
          },
          "description" : "Se false il testo indicato è inviato come semplice testo.",
          "default" : true
        },
        "notificationReading" : {
          "type" : "boolean",
          "example" : false,
          "xml" : {
            "name" : "confermaDiLetturaFlag"
          },
          "description" : "Se true si richiede una conferma dell'avvenuta apertura del messaggio.",
          "default" : false
        },
        "separateMailing" : {
          "type" : "boolean",
          "example" : false,
          "xml" : {
            "name" : "invioSeparatoFlag"
          },
          "description" : "Se true l'email viene inviata a gruppi di destinatari.",
          "default" : false
        }
      },
      "description" : "Richiesta per l'invio email",
      "xml" : {
        "name" : "richiestaSpedisciESalvaEmail"
      }
    },
    "SendAndSaveNewResponse" : {
      "type" : "object",
      "properties" : {
        "list" : {
          "type" : "array",
          "xml" : {
            "name" : "emailElaborata"
          },
          "description" : "Email processate nel sistema.",
          "items" : {
            "xml" : {
              "name" : "emailElaborata"
            },
            "$ref" : "#/definitions/EmailReferences"
          }
        },
        "attachments" : {
          "type" : "array",
          "xml" : {
            "name" : "allegato"
          },
          "description" : "File allegati all'email.",
          "items" : {
            "xml" : {
              "name" : "allegato"
            },
            "$ref" : "#/definitions/FileInfo"
          }
        },
        "attachmentsSize" : {
          "type" : "string",
          "example" : "821,77 KB",
          "xml" : {
            "name" : "dimensioneAllegati"
          },
          "description" : "Dimensione totale di tutti i file allegati all'email."
        },
        "idUtenteModPec" : {
          "type" : "string"
        },
        "mittentePEC" : {
          "type" : "boolean",
          "example" : true,
          "xml" : {
            "name" : "mittentePECFlag"
          },
          "description" : "Se true la casella di posta del mittente è PEC."
        },
        "notificationReading" : {
          "type" : "boolean",
          "example" : true,
          "xml" : {
            "name" : "confermaDiLetturaFlag"
          },
          "description" : "Se true è stata richiesta una conferma dell'avvenuta apertura del messaggio."
        }
      },
      "description" : "Risposta all'invio email",
      "xml" : {
        "name" : "rispostaSpedisciESalvaEmail"
      }
    },
...
  }
}